A saved world loads on Postgres 14+, where EXTRACT(EPOCH) is NUMERIC (#407)
## What

On PostgreSQL 14 and newer, the server panicked while loading a saved
world, so no dive could be created.

```
panicked at server\crates\meld-db\src\lib.rs:5775:42:
ColumnDecode { index: "\"updated_at_ms\"", source: "mismatched types; Rust type `f64` (as SQL type `FLOAT8`) is not compatible with SQL type `NUMERIC`" }
```

## Why

Since PostgreSQL 14, `EXTRACT(...)` returns `NUMERIC` instead of `double
precision`. `load_world` and `list_worlds` (#394/#397) decode
`updated_at_ms` as `f64`, so the decode fails. The Router calls
`list_worlds` at boot, so on any modern Postgres that already has a
saved world, the world task dies and every dive fails. Player data is
untouched; it just can't be reached.

## Fix

Cast to `float8` in both queries:

```sql
(EXTRACT(EPOCH FROM updated_at) * 1000)::float8 AS updated_at_ms
```

This works on both old and new Postgres. On versions before 14 the value
was already double precision, so the cast does nothing there.

## Testing

- Reproduced on PostgreSQL 17 (Windows) with an existing `worlds` row.
The server panicked at boot and dives never started.
- With the fix, the server boots and the world loads with no panic.
- Existing tests use the in-memory backend, which is why they didn't
catch this.

+++ server/crates/meld-db/src/lib.rs
@@ -5962,7 +5962,7 @@ impl Db {
             Backend::Pg(pool) => {
                 let row = sqlx::query(
                     "SELECT seed, tick_count, shift_generation, sections, delta,
-                            EXTRACT(EPOCH FROM updated_at) * 1000 AS updated_at_ms
+                            (EXTRACT(EPOCH FROM updated_at) * 1000)::float8 AS updated_at_ms
                      FROM worlds WHERE world_key = $1",
                 )
                 .bind(world_key)
@@ -5993,7 +5993,7 @@ impl Db {
             Backend::Pg(pool) => {
                 let rows = sqlx::query(
                     "SELECT world_key, seed, tick_count, shift_generation, sections, delta,
-                            EXTRACT(EPOCH FROM updated_at) * 1000 AS updated_at_ms
+                            (EXTRACT(EPOCH FROM updated_at) * 1000)::float8 AS updated_at_ms
                      FROM worlds ORDER BY world_key",
                 )
                 .fetch_all(pool)
